+++ commercetools-models/src/main/java/io/sphere/sdk/taxcategories/SubRate.java
@@ -1,7 +1,5 @@
 package io.sphere.sdk.taxcategories;
 
-import java.math.BigDecimal;
-
 /**
  * A SubRate is used to calculate the tax portions field in a cart or order.
  * It is useful if the total tax of a country is a combination of multiple taxes (e.g. state and local taxes)
@@ -9,9 +7,9 @@ import java.math.BigDecimal;
 public interface SubRate {
     String getName();
 
-    BigDecimal getAmount();
+    Double getAmount();
 
-    static SubRate of(final String name, final BigDecimal amount) {
+    static SubRate of(final String name, final Double amount) {
         return new SubRateImpl(name, amount);
     }
 }
